made the ngraph openvino implementation of the post training quantization by default
