+++ nncf/openvino/quantization/default_quantization.py
@@ -13,8 +13,7 @@
 
 from nncf.common.graph.operator_metatypes import UnknownMetatype
 from nncf.common.quantization.quantizer_propagation.structs import QuantizationTrait
-from nncf.experimental.openvino_native.graph.metatypes import openvino_metatypes as ov_metatypes
-
+from nncf.openvino.graph.metatypes import openvino_metatypes as ov_metatypes
 
 DEFAULT_OV_QUANT_TRAIT_TO_OP_DICT = {
     QuantizationTrait.INPUTS_QUANTIZABLE: [
