+++ tests/openvino/native/quantization/test_fq_configurations.py
@@ -10,8 +10,9 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 """
-from nncf.experimental.openvino_native.statistics.statistics import OVMinMaxTensorStatistic
+from nncf.openvino.statistics.statistics import OVMinMaxTensorStatistic
 from tests.post_training.test_calculate_quantizer_parameters import TemplateTestFQParams
+
 
 class TestFQParams(TemplateTestFQParams):
     @property
